requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
gcloud==0.18.3
googleapis-common-protos==1.69.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
oauth2client==4.1.3
packaging==24.2
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.1
Pyrebase4==4.8.0
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==1.26.20
